MAP-1402 LocationResourceResidentialTest for smaller files
